requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.1.2
django-rest-registration==0.3.12
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
mysql-connector-python==8.0.12
openapi-codec==1.3.2
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.7.1
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
py==1.6.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
